Archives: Expose the File and Directory classes to HLE
